requirements/base.txt
amqp==5.1.1
aniso8601==9.0.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
authtools==0.0.3
billiard==3.6.4.0
black==22.8.0
CacheControl==0.12.11
cachy==0.3.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
cleo==1.0.0a5
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
crashtest==0.3.1
Deprecated==1.2.13
distlib==0.3.6
Django==4.0.7
django-celery-beat==2.3.0
django-cors-headers==3.13.0
django-graphql-jwt==0.3.4
django-timezone-field==5.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
dulwich==0.20.45
filelock==3.8.0
flake8==5.0.4
django-graphql-jwt==0.3.4
graphene==3.1.1
graphene-django==2.0.0
graphql-core==3.2.1
graphql-relay==3.2.0
gunicorn==20.1.0
html5lib==1.1
idna==3.3
jaraco.classes==3.2.2
jsonschema==4.15.0
keyring==23.9.0
kombu==5.2.4
lockfile==0.12.2
mccabe==0.7.0
more-itertools==8.14.0
msgpack==1.0.4
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.1
pexpect==4.8.0
pkginfo==1.8.3
platformdirs==2.5.2
poetry==1.2.0
poetry-core==1.1.0
poetry-plugin-export==1.0.6
promise==2.3
prompt-toolkit==3.0.30
psycopg2==2.9.3
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
PyJWT==2.4.0
pylev==1.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
redis==4.3.4
requests==2.28.1
requests-toolbelt==0.9.1
Rx==1.6.1
shellingham==1.5.0
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
tomli==2.0.1
tomlkit==0.11.4
tzdata==2022.2
urllib3==1.26.12
vine==5.0.0
virtualenv==20.16.4
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.14.1
xattr==0.9.9